support plugin file preview
