+++ web/public/static/js/imjoy-preview.js
@@ -1,0 +1,8 @@
+if(window && window.location && window.location.protocol === 'file:'){
+    const preview = window.open("https://imjoy.io/#/preview");
+    window.addEventListener("message", function(ev) {
+        if (ev.data.type === "imjoy-app-ready") {
+            preview.postMessage({ type: "load-imjoy-plugin", code: document.body.innerHTML }, "*");
+        }
+    });
+}
