+++ web/src/router.js
@@ -3,7 +3,7 @@ import Router from "vue-router";
 import Imjoy from "@/components/Imjoy";
 import About from "@/components/About";
 import Home from "@/components/Home";
-
+import Preview from "@/components/Preview";
 Vue.use(Router);
 
 export const router_config = {
@@ -38,6 +38,11 @@ export const router_config = {
       path: "/about",
       name: "About",
       component: About,
+    },
+    {
+      path: "/preview",
+      name: "Preview",
+      component: Preview,
     },
     {
       path: "*",
